standard cards and better splitting of files
